Revert "create needed repositories for entities"
This reverts commit 0ef0c724eecf5439df328b031eaa4c71a69e9bb3.

+++ src/main/java/pl/sda/carrental/model/repository/DivisionRepository.java
@@ -7,8 +7,7 @@ import pl.sda.carrental.model.entity.Division;
 import java.util.List;
 
 @Repository
-public interface DivisionRepository extends JpaRepository<Division, Integer> {
-    
+public interface DivisionRepository extends JpaRepository<Division, Long> {
     @Override
     List<Division> findAll();
 }
